Code cleanup and formatting

+++ src/main/file/all/AllChannelPressureEvent.cpp
@@ -24,7 +24,7 @@ std::vector<char> AllChannelPressureEvent::mpcEventToBytes(
     std::vector<char> bytes(8);
 
     bytes[AllEvent::EVENT_ID_OFFSET] = AllEvent::CH_PRESSURE_ID;
-    AllEvent::writeTick(bytes, static_cast<int>(event->getTick()));
+    AllEvent::writeTick(bytes, event->getTick());
     bytes[AllEvent::TRACK_OFFSET] = static_cast<int8_t>(event->getTrack());
     bytes[AMOUNT_OFFSET] = static_cast<int8_t>(event->getAmount());
